+++ R/bsmote.R
@@ -203,13 +203,14 @@ step_bsmote_new <-
 prep.step_bsmote <- function(x, training, info = NULL, ...) {
   col_name <- recipes_eval_select(x$terms, training, info)
 
-  check_number_decimal(x$over_ratio, arg = "over_ratio", min = 0)
+  check_ratio(x$over_ratio, arg = "over_ratio")
   check_number_whole(x$neighbors, arg = "neighbors", min = 1)
   check_bool(x$all_neighbors, arg = "all_neighbors")
 
   check_1_selected(col_name)
   check_column_factor(training, col_name)
   warn_unused_levels(training, col_name)
+  check_ratio_column(x$over_ratio, training, col_name, arg = "over_ratio")
 
   recipes::check_name(
     tibble(x = logical(0)),
@@ -315,7 +316,8 @@ tunable.step_bsmote <- function(x, ...) {
     source = "recipe",
     component = "step_bsmote",
     component_id = x$id
-  )
+  ) |>
+    drop_per_class_ratio(x$over_ratio)
 }
 
 #' @rdname required_pkgs.step