+++ R/instance_hardness.R
@@ -195,12 +195,13 @@ step_instance_hardness_new <-
 prep.step_instance_hardness <- function(x, training, info = NULL, ...) {
   col_name <- recipes_eval_select(x$terms, training, info)
 
-  check_number_decimal(x$under_ratio, arg = "under_ratio", min = 0)
+  check_ratio(x$under_ratio, arg = "under_ratio")
   check_number_whole(x$neighbors, arg = "neighbors", min = 1)
 
   check_1_selected(col_name)
   check_column_factor(training, col_name)
   warn_unused_levels(training, col_name)
+  check_ratio_column(x$under_ratio, training, col_name, arg = "under_ratio")
 
   distance_cols <- recipes_argument_select(
     x$distance_with,
@@ -303,7 +304,8 @@ tunable.step_instance_hardness <- function(x, ...) {
     source = "recipe",
     component = "step_instance_hardness",
     component_id = x$id
-  )
+  ) |>
+    drop_per_class_ratio(x$under_ratio)
 }
 
 #' @rdname required_pkgs.step