+++ R/kmeans_smote.R
@@ -204,7 +204,7 @@ step_kmeans_smote_new <-
 prep.step_kmeans_smote <- function(x, training, info = NULL, ...) {
   col_name <- recipes_eval_select(x$terms, training, info)
 
-  check_number_decimal(x$over_ratio, arg = "over_ratio", min = 0)
+  check_ratio(x$over_ratio, arg = "over_ratio")
   check_number_whole(x$neighbors, arg = "neighbors", min = 1)
   check_number_whole(
     x$num_clusters,
@@ -228,6 +228,7 @@ prep.step_kmeans_smote <- function(x, training, info = NULL, ...) {
   check_1_selected(col_name)
   check_column_factor(training, col_name)
   warn_unused_levels(training, col_name)
+  check_ratio_column(x$over_ratio, training, col_name, arg = "over_ratio")
 
   recipes::check_name(
     tibble(x = logical(0)),
@@ -338,7 +339,8 @@ tunable.step_kmeans_smote <- function(x, ...) {
     source = "recipe",
     component = "step_kmeans_smote",
     component_id = x$id
-  )
+  ) |>
+    drop_per_class_ratio(x$over_ratio)
 }
 
 #' @rdname required_pkgs.step